# service/src/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mage API
  description: Mage API
  contact:
    name: Mage Support
    email: mage@nga.mil
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: 6.0.1
servers: []
security:
  - mageToken: []
...
paths:
...
  /auth/{authenticationStrategy}/signin:
    post:
      tags: [ Authentication & Authorization ]
      summary: Authentication
      description: Authenticate a user using a given strategy.
      operationId: authenticate
      security:
        []
      parameters:
        - name: authenticationStrategy
          in: path
          description: Authentication strategy.
          required: true
          schema:
            type: string
            enum: [ local, ldap, saml, oauth, openidconnect ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserIn' }
      responses:
        200:
          description: Successfully signed in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Authentication' }
...
  /api/events/{eventId}/observations/{observationId}:
    parameters:
      - $ref: '#/components/parameters/eventIdInPath'
      - $ref: '#/components/parameters/observationIdInPath'
    put:
      tags: [ Observation ]
      operationId: saveObservation
      description: >
        Create or update an observation with the specified observation ID for
        the specified event.  To create an observation, the requesting user
        must have `CREATE_OBSERVATION` permission and must be a member of a
        team with access to the event.  To update an observation, the
        requesting user must have `UPDATE_OBSERVATION_ALL` permission, or have
        `UPDATE_OBSERVATION_EVENT` permission as well as an ACL entry on the
        event with `read` permission, or be a member of a team with access to
        the event.
      requestBody:
        description: The observation properties
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationIn'
        required: true
      responses:
        200:
          description: Success - return the updated observation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
    get:
      tags: [ Observation ]
      operationId: getObservationById
      description: >
        Get the specified observation in the specified event.  The requesting
        user must have `READ_OBSERVATION_ALL` permission, or have
        `READ_OBSERVATION_EVENT` permission and an ACL entry on the event with
        `read` permission.
      parameters: &obsQueryParams
        - $ref: '#/components/parameters/observationQuery.fields'
        - $ref: '#/components/parameters/observationQuery.startDate'
        - $ref: '#/components/parameters/observationQuery.endDate'
        - $ref: '#/components/parameters/observationQuery.observationStartDate'
        - $ref: '#/components/parameters/observationQuery.observationEndDate'
        - $ref: '#/components/parameters/observationQuery.bbox'
        - $ref: '#/components/parameters/observationQuery.geometry'
        - $ref: '#/components/parameters/observationQuery.states'
        - $ref: '#/components/parameters/observationQuery.sort'
      responses:
        200:
          description: observation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
...
  /api/events/{eventId}/observations:
    parameters:
      - $ref: '#/components/parameters/eventIdInPath'
    get:
      tags: [ Observation ]
      description: >
        Get all the observations for the specified event that match the
        specified query parameters.  The requesting user must have
        `READ_OBSERVATION_ALL` permission, or have `READ_OBSERVATION_EVENT`
        permission and an ACL entry on the event with `read` permission.
      operationId: getObservationsForEvent
      # this used to be simply
      #   parameters: *obsQueryParams
      # referencing above, instead of the duplicated parameter array, but with
      # the update to js-yaml 4.1.x through openapi-enforcer, yaml anchors seem
      # to be broken
      parameters:
        - $ref: '#/components/parameters/observationQuery.fields'
        - $ref: '#/components/parameters/observationQuery.startDate'
        - $ref: '#/components/parameters/observationQuery.endDate'
        - $ref: '#/components/parameters/observationQuery.observationStartDate'
        - $ref: '#/components/parameters/observationQuery.observationEndDate'
        - $ref: '#/components/parameters/observationQuery.bbox'
        - $ref: '#/components/parameters/observationQuery.geometry'
        - $ref: '#/components/parameters/observationQuery.states'
        - $ref: '#/components/parameters/observationQuery.sort'
      responses:
        200:
          description: >
            Success - return the observation documents matching the query.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Observation'
  /api/events/{eventId}/observations/{observationId}/favorite:
    parameters:
      - $ref: '#/components/parameters/eventIdInPath'
      - $ref: '#/components/parameters/observationIdInPath'
    put:
      tags: [ Observation ]
      operationId: addObservationFavorite
      description: >
        Mark the specified observation as a favorite of the requesting user by
        adding the user's ID to the `favoriteUserIds` array on the observation.
        The requesting user must have `UPDATE_OBSERVATION_ALL` permission, or
        have `UPDATE_OBSERVATION_EVENT` permission and an ACL entry on the
        event with `read` permission.
      responses:
        200:
          description: Success - return the updated observation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
    delete:
      tags: [ Observation ]
      operationId: removeObservationFavorite
      description: >
        Remove the requesting user's ID from the `favoriteUserIds` of the
        specified observation.  The requesting user must have
        `UPDATE_OBSERVATION_ALL` permission, or have `UPDATE_OBSERVATION_EVENT`
        permission and an ACL entry on the event with `read` permission.
      responses:
        200:
          description: Success - return the updated observation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
  /api/events/{eventId}/observations/{observationId}/important:
    parameters:
      - $ref: '#/components/parameters/eventIdInPath'
      - $ref: '#/components/parameters/observationIdInPath'
    put:
      tags: [ Observation ]
      operationId: setObservationImportant
      description: >
        Mark the specified observation as important.  This operation allows
        administrators and event managers to flag important observations for
        all users with access to the event.  The requesting user must have
        `UPDATE_EVENT` permission and an ACL entry on the event with `update`
        permission.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationImportantCreate'
        required: true
      responses:
        200:
          description: Success - return the updated observation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
    delete:
      tags: [ Observation ]
      operationId: removeObservationImportant
      description: >
        Remove the important flag from the specified observation.  The
        requesting user must have `UPDATE_EVENT` permission and an ACL entry on
        the event with `update` permission.
      responses:
        200:
          description: Success - return the updated observation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observation'
  /api/events/{eventId}/observations/{observationId}/states:
    parameters:
      - $ref: '#/components/parameters/eventIdInPath'
      - $ref: '#/components/parameters/observationIdInPath'
    post:
      tags: [ Observation ]
      operationId: addObservationState
      description: >
        Set the current state of the specified observation.  The state is
        appended to a list of states on the observation, the head of which is
        the current state.  The requesting user must be the user that created
        the observation, have `UPDATE_EVENT` permission, or have an ACL entry on
        the event with `update` permission.
      requestBody:
        description: >
          The request body contains the name of the state to set on the
          target observation.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationStateCreate'
        required: true
      responses:
        201:
          description: Success - return the added state document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationState'
...
components:
  securitySchemes:
    authenticationToken:
      type: http
      description: >
        Token generated after successful authentication.  The client will exchange
        this token with an approved device for and API token.
      scheme: bearer
      bearerFormat: JWT
    mageToken:
      type: http
      description: >
        Mage generates an API token when a client authenticates
        initially.  The client can use the token for subsequent requests until
        it expires.
      scheme: bearer
  schemas:
...
    UserIn:
      type: object
      properties:
        username: { $ref: '#/components/schemas/User/properties/username' }
        displayName: { $ref: '#/components/schemas/User/properties/displayName' }
        email: { $ref: '#/components/schemas/User/properties/email' }
        phone:
          type: string
          format: phone
        roleId: { $ref: '#/components/schemas/User/properties/roleId' }
        iconMetadata:
          type: object
          properties:
            type: { $ref: '#/components/schemas/UserIcon/properties/type' }
            color: { $ref: '#/components/schemas/UserIcon/properties/color' }
            text: { $ref: '#/components/schemas/UserIcon/properties/text' }
        icon:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
        avatar:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
      allOf:
        - { $ref: '#/components/schemas/UserPasswordPair' }
...
    UserPasswordPair:
      type: object
      properties:
        password:
          type: string
        passwordconfirm:
          type: string
      oneOf:
        - { type: object, required: [ password, passwordconfirm ] }
        - { type: object, not: { type: object, required: [ password, passwordconfirm ] } }
    Authentication:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/UserRef'
        token:
          type: string
...
    ObservationIn:
      description: >
        This schema defines the observation properties a client sends to save
        a new or updated observation.
      allOf:
        - $ref: 'geojson.yaml#/definitions/feature'
        - type: object
          properties:
            properties:
              $ref: '#/components/schemas/ObservationFeatureProperties'
        - type: object
          required: [ type, geometry, properties ]
    Observation:
      title: Observation
      description: >
        The `Observation` is the operative concept of Mage's geospatial data
        collection functionality.  Observation documents are the geo-tagged
        form and media records that Mage users submit to the Mage server.
        Observation documents are extensions of GeoJSON feature documents, so
        all observations have a `type` property whose value is `Feature`, as
        well as a `geometry` property that holds the geospatial geometry of the
        observation, such as point or polygon coordinates.  The values of the
        observation's form fields reside under the GeoJSON `properties` key.
      type: object
      properties:
        id:
          type: string
          description: The unique observation ID
        createdAt:
          type: string
          format: date-time
        lastModified:
          type: string
          format: date-time
        userId:
          $ref: '#/components/schemas/User/properties/id'
        deviceId:
          $ref: '#/components/schemas/Device/properties/id'
        eventId:
          $ref: '#/components/schemas/Event/properties/id'
        state:
          $ref: '#/components/schemas/ObservationState'
        url:
          type: string
          format: uri
          description: >
            This is the absolute URL of the observation where the client can
            fetch and update the observation document.
        favoriteUserIds:
          type: array
          items:
            $ref: '#/components/schemas/User/properties/id'
        attachements:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
      allOf:
        - $ref: '#/components/schemas/ObservationIn'
      example: {
        "createdAt": "2019-09-25T19:50:05.065Z",
        "deviceId": "5d697f88951c51a926e9dba4",
        "geometry": {
          "type": "Point",
          "coordinates": [
            -0.08910298347473145,
            51.527068718503244
          ],
        },
        "lastModified": "2019-10-01T04:24:45.466Z",
        "properties": {
          "timestamp": "2019-09-25T19:48:55.859Z",
          "forms": [
            {
              "formId": 3,
              "field1": [
                "Umbarela"
              ],
            }
          ],
        },
        "type": "Feature",
        "userId": "5d697f88b8ed5f29bef7b10c",
        "favoriteUserIds": [
          "5d697f88b8ed5f29bef7b10c"
        ],
        "important": {
          "description": "Come back later",
          "timestamp": "2019-10-01T04:24:45.466Z",
          "userId": "5d697f88b8ed5f29bef7b10c"
        },
        "attachments": [
          {
            "contentType": "image/jpeg",
            "size": 24875,
            "name": "c1816efcab6e00c39623ddd8262b852e.jpg",
            "relativePath": "observations2/2019/9/25/c1816efcab6e00c39623ddd8262b852e.jpg",
            "lastModified": "2019-09-25T19:50:22.383Z",
            "oriented": false,
            "id": "5d8bc4fe038559417fffbb55",
            "url": "http://localhost:4242/api/events/2/observations/5d8bc4ed038559417fffbb53/attachments/5d8bc4fe038559417fffbb55"
          }
        ],
        "id": "5d8bc4ed038559417fffbb53",
        "eventId": 2,
        "url": "http://localhost:4242/api/events/2/observations/5d8bc4ed038559417fffbb53",
        "state": {
          "userId": "5d697f88b8ed5f29bef7b10c",
          "name": "active",
          "id": "5d8bc4ed038559417fffbb54",
          "url": "http://localhost:4242/api/events/2/observations/5d8bc4ed038559417fffbb53/states/5d8bc4ed038559417fffbb54"
        }
      }
...
    ObservationImportantCreate:
      allOf:
        - $ref: '#/components/schemas/ObservationImportant'
        - type: object
          required: [ description ]
    ObservationState:
      type: object
      properties:
        id:
          type: string
        name:
          $ref: '#/components/schemas/ObservationStateName'
        userId:
          $ref: '#/components/schemas/User/properties/id'
          description: >
            The ID of the user that set this state on the parent observation.
    ObservationStateCreate:
      allOf:
        - $ref: '#/components/schemas/ObservationState'
        - type: object
          required: [ name ]
...
  parameters:
...
    eventIdInPath:
      in: path
      name: eventId
      description: The ID of the target event document
      required: true
      example: 1234
      schema: { $ref: '#/components/schemas/Event/properties/id' }
...
    observationIdInPath:
      in: path
      name: observationId
      description: The ID of the target observation document
      required: true
      schema: { $ref: '#/components/schemas/Observation/properties/id' }
...
    observationQuery.fields:
      in: query
      name: fields
      description: >
        The form fields to project in the result observation documents (JSON)
      explode: false
      schema:
        type: array
        items:
          type: string
    observationQuery.startDate:
      in: query
      name: startDate
      schema:
        type: string
        format: date-time
      description: >
        The low end of the range for the observations' `lastModified`
        property
    observationQuery.endDate:
      in: query
      name: endDate
      schema:
        type: string
        format: date-time
      description: >
        The high end of the range for the observations' `lastModified`
        property
    observationQuery.observationStartDate:
      in: query
      name: observationStartDate
      schema:
        type: string
        format: date-time
      description: >
        The low end of the range for the observations' `timestamp`
        property
    observationQuery.observationEndDate:
      in: query
      name: observationEndDate
      schema:
        type: string
        format: date-time
      description: >
        The low end of the range for the observations' `lastModified`
        property
    observationQuery.bbox:
      in: query
      name: bbox
      description: >
        Bounding box array as a JSON string, e.g.
        `[xMin, yMin, xMax, yMax]`
      explode: false
      schema:
        $ref: 'geojson.yaml#/definitions/boundingBox'
    observationQuery.geometry:
      in: query
      name: geometry
      description: >
        A URL-encoded, stringified JSON object that is a GeoJSON geometry
        as defined in geojson.yaml#/definitions/geometryObject
      schema:
        type: string
        format: json
    observationQuery.states:
      in: query
      name: states
      explode: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/ObservationStateName'
    observationQuery.sort:
      in: query
      name: sort
      description: >
        The `sort` query parameter is a comma-separated list of sort keys.
        Currently, this operation only supports sorting on the
        `lastModified` key.
      explode: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/SortKey'
...